implant-hub-fixed/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
psycopg2-binary==2.9.9
pydantic-settings==2.6.1
